requirements.txt
Flask==2.3.0
google-cloud-storage>=2.0.0,<3.0.0
google-generativeai==0.8.3
